+++ example/trigger/src/main/java/org/apache/iotdb/trigger/AlertingExample.java
@@ -21,9 +21,9 @@ package org.apache.iotdb.trigger;
 
 import org.apache.iotdb.db.engine.trigger.api.Trigger;
 import org.apache.iotdb.db.engine.trigger.api.TriggerAttributes;
-import org.apache.iotdb.db.sink.alertmanager.AlertManagerConfiguration;
-import org.apache.iotdb.db.sink.alertmanager.AlertManagerEvent;
-import org.apache.iotdb.db.sink.alertmanager.AlertManagerHandler;
+import org.apache.iotdb.db.engine.trigger.sink.alertmanager.AlertManagerConfiguration;
+import org.apache.iotdb.db.engine.trigger.sink.alertmanager.AlertManagerEvent;
+import org.apache.iotdb.db.engine.trigger.sink.alertmanager.AlertManagerHandler;
 
 import java.io.IOException;
 import java.util.HashMap;
